move compiled payload into a new header. add docstrings for payload classes
