+++ include/firestarter/Environment/X86/Payload/CompiledX86Payload.hpp
@@ -22,25 +22,35 @@
 #pragma once
 
 #include "asmjit/core/jitruntime.h"
-#include "firestarter/Environment/Payload/Payload.hpp"
+#include "firestarter/Environment/Payload/CompiledPayload.hpp"
 #include "firestarter/Logging/Log.hpp"
 #include <memory>
 
 namespace firestarter::environment::x86::payload {
 
+/// This class provides the functionality to compile a payload created with asmjit and create a unique pointer to the
+/// CompiledPayload class which can be used to execute the functions of this payload.
 class CompiledX86Payload final : public environment::payload::CompiledPayload {
 private:
   // NOLINTNEXTLINE(cppcoreguidelines-avoid-non-const-global-variables)
   inline static asmjit::JitRuntime Runtime = asmjit::JitRuntime();
 
+  /// Custom deleter to release the memory of the high load function from the asmjit runtime.
+  /// \arg Payload The pointer to this class
   static void deleter(CompiledX86Payload* Payload) {
     if (Payload && Payload->highLoadFunctionPtr()) {
       Runtime.release(Payload->highLoadFunctionPtr());
     }
   }
-
+  /// Custom deleter to release the memory of the high load function from the asmjit runtime.
+  /// \arg Payload The pointer to this class
   static void deleter(CompiledPayload* Payload) { deleter(dynamic_cast<CompiledX86Payload*>(Payload)); }
 
+  /// Wrap the CompiledPayload class and forward all arguments.
+  /// \arg Stats The stats of the high load function from the payload.
+  /// \arg PayloadPtr A unique pointer to the payload class to allow calling the init and low load functions which do
+  /// not change based on different payload settings.
+  /// \arg HighLoadFunction The pointer to the compiled high load function.
   CompiledX86Payload(const environment::payload::PayloadStats& Stats,
                      std::unique_ptr<environment::payload::Payload>&& PayloadPtr, HighLoadFunctionPtr HighLoadFunction)
       : CompiledPayload(Stats, std::move(PayloadPtr), HighLoadFunction) {}
@@ -49,6 +59,12 @@ public:
   CompiledX86Payload() = delete;
   ~CompiledX86Payload() override = default;
 
+  /// Create a unique pointer to a compiled payload from payload stats and assembly in a code holder.
+  /// \targ DerivedPayload The payload class from which the CodeHolder with the assembly was created from.
+  /// \arg Stats The stats of the payload that is contained in the CodeHolder.
+  /// \arg Code The CodeHolder that contains the assembly instruction making up the payload. This will be added to the
+  /// JitRuntime and a pointer to the function will be provided to the CompiledPayload class.
+  /// \returns The unique pointer to the compiled payload.
   template <class DerivedPayload>
   [[nodiscard]] static auto create(environment::payload::PayloadStats Stats, asmjit::CodeHolder& Code) -> UniquePtr {
     HighLoadFunctionPtr HighLoadFunction{};
